+++ core/src/discrete_wavelet_transform.c
@@ -3,19 +3,19 @@
 #include <string.h>
 #include <stdio.h>
 
-static inline int32_t round_int(int32_t numerator, int32_t denominator) {
+static inline int16_t round_int(int32_t numerator, int32_t denominator) {
     return (numerator) < 0 ? ((numerator) - ((denominator)/2)) / (denominator) : ((numerator) + ((denominator)/2)) / (denominator);
 }
 
-static void forward_DWT(int32_t* data, size_t width) {
+static void forward_DWT(int16_t* data, size_t width) {
 
 	//cache line for in place operation
-	int32_t cache[width];
-	memcpy(&cache, data, width * sizeof(int32_t));
+	int16_t cache[width];
+	memcpy(&cache, data, width * sizeof(int16_t));
 
 	uint32_t n = width >> 1; //number of coefficients in pass
-	int32_t* highpass = &data[n];
-	int32_t* lowpass = &data[0];
+	int16_t* highpass = &data[n];
+	int16_t* lowpass = &data[0];
 
     highpass[0] = cache[1] - round_int((9 * (cache[0] + cache[2]) - (cache[2] + cache[4])), 16);
     lowpass[0] = cache[0] - round_int(-highpass[0], 2);
@@ -26,20 +26,19 @@ static void forward_DWT(int32_t* data, size_t width) {
 	for(size_t i = 1; i < n - 2; ++i) {
         highpass[i] = cache[2*i+1] - round_int((9 * (cache[2*i] + cache[2*i+2]) - (cache[2*i-2] + cache[2*i+4])), 16);
         lowpass[i] = cache[2*i] - round_int(-(highpass[i-1] + highpass[i]), 4);
-        
 	}
 
     lowpass[n-2] = cache[2*n-4] - round_int(-(highpass[n-3] + highpass[n-2]), 4);
     lowpass[n-1] = cache[2*n-2] - round_int(-(highpass[n-2] + highpass[n-1]), 4);
 }
 
-static void backward_DWT(int32_t* data, size_t width) {
+static void backward_DWT(int16_t* data, size_t width) {
 
     size_t n = width >> 1;
-    int32_t cache[width];
-	int32_t* lowpass = &cache[0];
-	int32_t* highpass = &cache[n];
-    memcpy(&cache, data, width * sizeof(int32_t));
+    int16_t cache[width];
+	int16_t* lowpass = &cache[0];
+	int16_t* highpass = &cache[n];
+    memcpy(&cache, data, width * sizeof(int16_t));
     
     data[0] = lowpass[0] + round_int(-highpass[0], 2);
 
@@ -57,7 +56,7 @@ static void backward_DWT(int32_t* data, size_t width) {
     data[2*n-1] = highpass[n-1] + round_int(9 * data[2*n-2] - data[2*n-4], 8);
 }
 
-void discrete_wavelet_transform_2D(int32_t* data, size_t data_width, size_t data_height, uint8_t transform_levels, uint8_t inverse_flag) {
+void discrete_wavelet_transform_2D(int16_t* data, size_t data_width, size_t data_height, uint8_t transform_levels, uint8_t inverse_flag) {
 
     if(inverse_flag) {
         for(int8_t level = transform_levels-1; level >= 0; --level) {
@@ -65,7 +64,7 @@ void discrete_wavelet_transform_2D(int32_t* data, size_t data_width, size_t data
             uint32_t current_height = data_height >> level;
 
             //Vertical
-			int32_t temp_column[current_height];
+			int16_t temp_column[current_height];
             for(uint32_t column = 0; column < current_width; ++column) {
 
 				//Copy current column to temp_column
@@ -74,7 +73,7 @@ void discrete_wavelet_transform_2D(int32_t* data, size_t data_width, size_t data
 					temp_column[row] = data[row * data_width + column];
 				}
 
-				backward_DWT((int32_t*)&temp_column, current_height);
+				backward_DWT((int16_t*)&temp_column, current_height);
 
 				for(uint32_t row = 0;  row < current_height; ++row) {
 					data[row * data_width + column] = temp_column[row];
@@ -82,7 +81,7 @@ void discrete_wavelet_transform_2D(int32_t* data, size_t data_width, size_t data
 			}
 
             //Horizontal
-			int32_t temp_row[current_width];
+			int16_t temp_row[current_width];
 			//Iterate each row
             for(uint32_t row = 0; row < current_height; ++row) {
 
@@ -91,7 +90,7 @@ void discrete_wavelet_transform_2D(int32_t* data, size_t data_width, size_t data
 					temp_row[column] = data[row * data_width + column];
 				}
 
-                backward_DWT((int32_t*)&temp_row, current_width);
+                backward_DWT((int16_t*)&temp_row, current_width);
 
 				for(uint32_t column = 0;  column < current_width; ++column) {
 					data[row * data_width + column] = temp_row[column];
@@ -108,7 +107,7 @@ void discrete_wavelet_transform_2D(int32_t* data, size_t data_width, size_t data
 
 		
 		//Horizontal
-		int32_t temp_row[current_width];
+		int16_t temp_row[current_width];
 		//Iterate each row
 		for(uint32_t row = 0; row < current_height; ++row) {
 
@@ -117,7 +116,7 @@ void discrete_wavelet_transform_2D(int32_t* data, size_t data_width, size_t data
 				temp_row[column] = data[row * data_width + column];
 			}
 
-			forward_DWT((int32_t*)&temp_row, current_width);
+			forward_DWT((int16_t*)&temp_row, current_width);
 
 			for(uint32_t column = 0;  column < current_width; ++column) {
 				data[row * data_width + column] = temp_row[column];
@@ -125,7 +124,7 @@ void discrete_wavelet_transform_2D(int32_t* data, size_t data_width, size_t data
 		}
 
 		//Vertical
-		int32_t temp_column[current_height];
+		int16_t temp_column[current_height];
 		for(uint32_t column = 0; column < current_width; ++column) {
 
 			//Copy current column to temp_column
@@ -134,7 +133,7 @@ void discrete_wavelet_transform_2D(int32_t* data, size_t data_width, size_t data
 				temp_column[row] = data[row * data_width + column];
 			}
 
-			forward_DWT((int32_t*)&temp_column, current_height);
+			forward_DWT((int16_t*)&temp_column, current_height);
 
 			for(uint32_t row = 0;  row < current_height; ++row) {
 				data[row * data_width + column] = temp_column[row];
